+++ packages/kit/src/runtime/server/constants.js
@@ -1,0 +1,1 @@
+export const NULL_BODY_STATUS = [101, 103, 204, 205, 304];