+++ packages/kit/test/apps/basics/src/routes/load/fetch-body-stream-b64/data/+server.js
@@ -1,0 +1,7 @@
+export const GET = () => {
+	return new Response(new Uint8Array([1, 2, 3, 4]));
+};
+
+export const POST = async ({ request }) => {
+	return new Response(await request.arrayBuffer());
+};